Move issue constants to own file
